Deleted: 'AgentID' Renamed: 'EntityID' -> 'AgentID'
There were some problems caused by this change so I'm sorry to everyone
for the disruption regarding my 'EntityID' shenanigans, it was a
short-sighted move on my part to mess with it. The change in this commit
was done because the statistics team wants to be able to have AgentID's
and PrototypeID's that cannot be compared or casted, and there would be too
much disunity if there was a third distinct class 'EntityID' that the other
two ID classes would not be able to extend (for reasons mentioned above).