Current ID setup:
discrete classes (no relationships):
- AgentID
- EntityID
- PrototypeID

important methods:
- Agent.getAgentID()
- Prototype.getPrototypeID()
- Entity.getEntityID()

Each ID class implements 'compareTo' and 'equals'.
In the next commit I will modify the 'toString' methods for these
classes to prefix the id number with the type of id and a colon
as a delimiter.

An Agent will have an EntityID as a subclass of Entity, as well
as an AgentID.

A Prototype will have an EntityID as a subclass of Entity, as well
as a PrototypeID.