Changed expression functions a little bit, added a numArgs() method that returns the number of arguments a function takes
Also for functions some were like a(this, b, c,d) and some were like b(c, d, e)
The 'this' was unnecessary, so I took out all of the this's that pointed to itself and hard coded it, as its better to have everything have the same format.
Couldn't test because GUI is broken.
Should work unless I missed one of the 'this' in the conways or RPS